payment/feat-176 history 생성 코드 간소화. payment 이벤트 소비 코드 추가
