+++ order/src/main/java/com/ll/order/domain/service/order/create/AbstractOrderCreationService.java
@@ -8,7 +8,6 @@ import com.ll.order.domain.client.UserServiceClient;
 import com.ll.order.domain.exception.OrderErrorCode;
 import com.ll.order.domain.model.entity.Order;
 import com.ll.order.domain.model.entity.OrderItem;
-import com.ll.order.domain.model.entity.history.OrderHistoryBuilder;
 import com.ll.order.domain.model.entity.history.OrderHistoryEntity;
 import com.ll.order.domain.model.enums.order.OrderHistoryActionType;
 import com.ll.order.domain.model.enums.order.OrderStatus;
@@ -166,15 +165,17 @@ public abstract class AbstractOrderCreationService {
                     // 주문 상태 변경 이력 저장 (재고 차감 실패)
                     String errorMessage = String.format("재고 차감 실패 - productCode: %s, error: %s",
                             orderItem.getProductCode(), e.getMessage());
-                    OrderHistoryEntity failHistory = OrderHistoryBuilder.builder()
-                            .order(order)
-                            .orderItems(orderItems)
-                            .actionType(OrderHistoryActionType.STATUS_CHANGE)
-                            .previousStatus(previousStatus)
-                            .reason("재고 차감 실패")
-                            .errorMessage(errorMessage)
-                            .createdBy("SYSTEM")
-                            .build();
+                    OrderHistoryEntity failHistory = OrderHistoryEntity.create(
+                            order,
+                            orderItems,
+                            OrderHistoryActionType.STATUS_CHANGE,
+                            previousStatus,
+                            "재고 차감 실패",
+                            errorMessage,
+                            null,
+                            null,
+                            "SYSTEM"
+                    );
                     orderHistoryJpaRepository.save(failHistory);
                 }
             }
