+++ order/src/main/java/com/ll/order/domain/service/order/create/strategy/CartOrderCreationStrategy.java
@@ -8,7 +8,6 @@ import com.ll.order.domain.client.UserServiceClient;
 import com.ll.order.domain.exception.OrderErrorCode;
 import com.ll.order.domain.model.entity.Order;
 import com.ll.order.domain.model.entity.OrderItem;
-import com.ll.order.domain.model.entity.history.OrderHistoryBuilder;
 import com.ll.order.domain.model.entity.history.OrderHistoryEntity;
 import com.ll.order.domain.model.enums.order.OrderStatus;
 import com.ll.order.domain.model.enums.payment.PaidType;
@@ -118,7 +117,7 @@ public class CartOrderCreationStrategy extends AbstractOrderCreationService {
         orderItemJpaRepository.saveAll(orderItems);
 
         // 주문 생성 이력 저장
-        OrderHistoryEntity orderHistory = OrderHistoryBuilder.createOrderHistory(savedOrder, orderItems);
+        OrderHistoryEntity orderHistory = OrderHistoryEntity.createOrderHistory(savedOrder, orderItems);
         orderHistoryJpaRepository.save(orderHistory);
 
         return new OrderCreationResult(savedOrder, orderItems);
@@ -147,7 +146,7 @@ public class CartOrderCreationStrategy extends AbstractOrderCreationService {
             orderJpaRepository.save(order);
 
             // 주문 상태 변경 이력 저장 (결제 성공)
-            OrderHistoryEntity successHistory = OrderHistoryBuilder.createPaymentSuccessHistory(
+            OrderHistoryEntity successHistory = OrderHistoryEntity.createPaymentSuccessHistory(
                     order, orderItems, previousStatus, "예치금");
             orderHistoryJpaRepository.save(successHistory);
 
@@ -162,7 +161,7 @@ public class CartOrderCreationStrategy extends AbstractOrderCreationService {
             orderJpaRepository.save(order);
 
             // 주문 상태 변경 이력 저장 (결제 실패)
-            OrderHistoryEntity failHistory = OrderHistoryBuilder.createPaymentFailHistory(
+            OrderHistoryEntity failHistory = OrderHistoryEntity.createPaymentFailHistory(
                     order, orderItems, order.getOrderStatus(), "예치금", e.getMessage());
             orderHistoryJpaRepository.save(failHistory);
